@@ -2,8 +2,11 @@
   "name": "ts-fsrs",
   "version": "3.0.4",
   "description": "ts-fsrs is a TypeScript package used to implement the Free Spaced Repetition Scheduler (FSRS) algorithm. It helps developers apply FSRS to their flashcard applications, thereby improving the user learning experience.",
-  "main": "dist/index.js",
+  "main": "dist/ts-fsrs.js",
+  "module": "dist/ts-fsrs.mjs",
+  "browser": "dist/ts-fsrs.umd.js",
   "types": "dist/index.d.ts",
+  "type": "module",
   "keywords": [
     "SuperMemo",
     "Anki",
@@ -14,33 +17,49 @@
     "seedrandom": "^3.0.5"
   },
   "devDependencies": {
-    "@types/jest": "^29.5.5",
-    "@types/node": "^20.6.2",
-    "@types/seedrandom": "^3.0.5",
-    "@typescript-eslint/eslint-plugin": "^6.7.0",
-    "@typescript-eslint/parser": "^6.7.0",
-    "eslint": "^8.44.0",
+    "@babel/core": "^7.23.3",
+    "@babel/preset-env": "^7.23.3",
+    "@rollup/plugin-alias": "^5.0.1",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-json": "^6.0.1",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@types/jest": "^29.5.8",
+    "@types/node": "^20.9.1",
+    "@types/seedrandom": "^3.0.8",
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "eslint": "^8.53.0",
     "eslint-config-prettier": "^9.0.0",
     "jest": "^29.7.0",
-    "prettier": "^3.0.3",
+    "prettier": "^3.1.0",
+    "rimraf": "^5.0.5",
+    "rollup": "^4.4.1",
+    "rollup-plugin-typescript2": "^0.36.0",
     "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.1.6"
+    "tslib": "^2.6.2",
+    "typedoc": "^0.25.3",
+    "typescript": "^5.2.2"
   },
   "scripts": {
     "lint": "eslint --fix src/ && prettier --write src/",
-    "dev": "ts-node src/fsrs/index.ts",
+    "dev": "rollup -c -w",
     "test": "jest --passWithNoTests",
-    "build": "tsc --project ./tsconfig.json",
+    "prebuild": "rimraf ./dist",
+    "build": "rollup --c ",
+    "build:types": "tsc --project ./tsconfig.json --declaration true",
     "major": "npm version major",
     "minor": "npm version minor",
     "patch": "npm version patch",
-    "test_publish": "yalc publish"
+    "test_publish": "yalc publish",
+    "predocs": "rimraf ./docs",
+    "docs": "typedoc src/fsrs --out docs",
+    "clean": "rimraf ./dist ./docs"
   },
   "author": "ishiko",
   "license": "MIT",
   "files": [
-    "dist/**/*",
+    "dist",
     "README.md",
     "LICENSE"
   ],
